CDPD-25078: Multirow insert with decimals needs to use Impala common decimal type.
When there are multiple rows in an array, the Impala code is called
to check for the compatible type.

Specifically, there was a bug on the following insert which uses arrays:

insert overwrite values_test_float_tbl values (1000000), (1.1), (0.07), (33.333);

The insert statement needed common datatypes to promote to across the rows as
defined by Impala semantics.

Change-Id: Ib6ef45031dbdbebc36db0412431752eaa41c28a6