CDPD-55112: HIVE-24957: HIVE-24999: Backport: Inefficient & wrong CBO plans in the presence of subqueries (Stamatis Zampetakis, reviewed by Krisztian Kasa)
* HIVE-24999: HiveSubQueryRemoveRule generates invalid plan for IN subquery with correlations

1. Add workaround for CALCITE-4574 in HiveRelBuilder to avoid generating
invalid plans (filter with references to columns which do not exist).

2. Adapt HiveRelDecorrelator based on new plans generated by HiveSubQueryRemoveRule

2a. Remove workaround getNewForOldInputRef that was needed due to the
invalid plans.

2b. Adapt input references based on new the input operator (frame) inside
decorrelateInputWithValueGenerator method.

3c. Refactor DecorrelateRexShuttle#visitCall to improve readability and
cover a few more corner cases.

3. Add subquery_in_invalid_intermediate_plan.q with problematic plans
relevant for the case.

4. Add CBO explain plans in queries related to masking since there are
easier to read and compare with. There are few plan regressions that
will be fixed by HIVE-24957.

* HIVE-24957: Wrong results when subquery has COALESCE in correlation predicate

1. Add plan transformations before starting the core RelDecorrelator logic
to bring the plan into an equivalent but more convenient form that can be
decorrelated into more efficient and correct plans.

2. Adapt HiveRelDecorrelator#decorrelateInputWithValueGenerator to avoid
creating value generator for already satisfied correlations present in
the input.

3. Based on the changes above many plans with subqueries become more
efficient since the value generator is no longer necessary and it is dropped.

4. Add subquery_complex_correlation_predicates.q which includes queries
generating wrong results without the new transformations.

5. Add CBO plans in few queries since they are easier to read and reason
about correctness and efficiency.

Change-Id: Iae778d47f15438085049641d85bac2a6338854e3