CDPD-84192: HIVE-25046: Log CBO plans right after major transformations (Stamatis Zampetakis, reviewed by Jesus Camacho Rodriguez)
Change to avoid the following shortcomings:
1. If there is a failure (exception) in some transformation step
intermediate plans are lost, thus losing track of which plan led to the
problem.
2. Intermediate logs are useful for identifying plan problems while
working on a patch; unfortunately the logs are explicitly disabled in
test mode which means that in order to appear the respective code needs
to change every time we need to see those logs.
3. Logging at the end necessitates keeping additional local variables
which makes the code slightly harder to read.

It also adds some additional logging which was not present at all
before.

Closes apache/hive#2205

Change-Id: I66df76e525c405102923d627003b869118bd84cf