CDPD-44520 HIVE-25093: date_format() UDF is returning output in UTC time zone only (Ashish Sharma, reviewed by Adesh Rao, Nikhil Gupta, Sankar Hariappan)
Signed-off-by: Sankar Hariappan <sankarh@apache.org>
Closes (# 2252)

(cherry-picked from commit 88048e7d2c35e9046679131e4dab3bde73bdef87)

TestGenericUDFDateFormat#testJulianDates fails in this commit although
it is marked with @Ignore. Probably this is due to surefire/junit
integration problem but since the test will be fixed and re-enabled in
the next commit (HIVE-25268) we are leaving it as is for now.

Change-Id: I5f941466d3a2b865b90dd6d6a5b17266440cd831