CDPD-27892: Add SELECT, INSERT support for feng-kudu
This patch is based on the patch for the FENG branch at
https://gerrit.sjc.cloudera.com/c/cdh/hive/+/135750. On top of the patch
for the FENG branch, other than resolving various conflicts when
backporting the patch to cdw-master, we had to implement the following
changes to resolve various issues.

1. We corrected a mistake in ImpalaHdfsScanRel.java. In particular, in
the original patch, the author changed the name of the variable that is
to be returned upon the completion of ImpalaHdfsScanRel#getPlanNode()
from "hdfsScanNode" to "impalaScanNode" by introducing 'impalaScanNode'
as a local variable to reflect the fact that the returned scan node
could be an ImpalaKuduScanNode as well. Recall that 'hdfsScanNode' was a
field under the class of ImpalaHdfsScanRel and thus returning that newly
introduced local variable was incorrect. This patch corrected this by
changing the field name from "hdfsScanNode" to "impalaScanNode".

2. In SemanticAnalyzer#checkInsertDestinationTableFormat(), we
additionally added KuduInputFormat as a supported table type for the
INSERT statement.

3. In TargetHelper#createDynPartitionColPositions(), we additionally
added ImpalaKuduTable as a possible instance to allow insertion into a
Kudu table.

Testing:
 - Verified that the patch could pass TPCH queries against Kudu tables.
 - Briefly verified that a user could insert rows into a Kudu table
   created by the classic Impala. The inserted rows could be shown in
   both beeline and impala-shell.

Change-Id: I384d213faad3c2f85bb3c96a15346c95037ab415