CDPD-26879: Avoid discrepancies between Impala and FENG BuiltinsDB metadata
Previously, a static version of the Impala functions were stored in
a couple of json files. When new functions were added into Impala,
the json files had to be manually synced.

The Hive Impala structures will now be populated directly from the
BuiltinsDb.

The json file is still needed for SQL functions that exist in Calcite
and are supported, but do not exist directly in Impala. One example
of this is the 'between' operator which gets converted to a > and a
< operator.

The json file also previously stored some extra information needed on
the Calcite side. One example of this is function names that do not
exactly match, like "ne" in Impala maps to "!=" in Calcite. These
mappings are stored in the FunctionDetailStatics class.

The test framework structure also requires json files because the
Impala ".so" file is not available when the qtests are run. Some
FunctionWrapper interfaces exist to allow the test framework to
emulate what the ScalarFunction and AggregateFunction classes provide.

Change-Id: I57a5ee6281842b525bc6b4ddd4bbb082f73957f0