CDPD-28215: Change how ImpalaHdfsPartition is initialized after IMPALA-10502
IMPALA-10502 introduced a new input argument of "createEventId" for a
constructor of HdfsPartition. ImpalaHdfsPartition inherits HdfsPartition
and invokes this constructor to initialize its fields. Due to the change
to the signature of the constructor of HdfsPartition, we have to add the
missing argument when the constructor of HdfsPartition is called.

Specifically, there are 2 constructors provided by HdfsPartition. Before
IMPALA-10502, those constructors took in 18 and 19 input arguments,
respectively. IMPALA-10502 added one input argument for the latter
constructor and thus those constructors took in 18 and 20 input
arguments after IMPALA-10502. Recall that ImpalaHdfsPartition called the
constructor of 19 input arguments before IMPALA-10502 and thus after
IMPALA-10502 there was no constructor with compatible signature, causing
the compilation error for ImpalaHdfsPartition.

Testing:
 - Verified locally that Hive could be compiled after this patch.

Change-Id: I02e1b6774f8afac2bbafb73b8d8a6bbf08e74f38