CDPD-27938: Add CREATE, DROP support for feng-kudu
This patch is based on the patch for the FENG branch at
https://gerrit.sjc.cloudera.com/c/cdh/hive/+/140757. On top of the patch
for the FENG branch, we had to implement the following changes to
resolve various issues.

1. We added methods to EngineRuntimeHelper.java and implemented them in
NativeEngineRuntimeHelper.java and ImpalaRuntimeHelper.java. This way
the code in the ql module will be able to access the methods under the
impala module. Without this change, this patch could not be compiled.

2. We changed the condition based on which we determine whether a table
should be created on the Kudu server in CreateTableOperation.java.
Without this change, the CTAS statement could not create a Kudu table on
the Kudu server when necessary.

3. We added a null check in TaskCompiler.java to make sure the
LoadFileDesc is not null before verifying whether 'serName' equals
'KUDU_SERDE' to avoid a NullPointerException.

4. We disabled the tests under various sub-modules of hcatalog against
the newly registered storage format descriptor of
KuduStorageFormatDescriptor in
org.apache.hadoop.hive.ql.io.StorageFormatDescriptor. Those tests are
not supported because when the backend execution engine is TEZ, Hive
does not support creating Kudu tables on the Kudu server, which is a
prerequisite for those tests to run. It should be fine to disable them
considering the fact that hcatalog is not currently supported and that
KuduStorageFormatDescriptor is only used in the context of Unified
Analytics.

On the other hand, we also corrected a mistake in CDPD-27892.
Specifically, for a CTAS statement to create a Kudu table, 'impalaTable'
in TargetHelper#createDynPartitionColPositions() is an instance of
KuduTable, whereas for an INSERT statement, it is an instance of
ImpalaKuduTable, which inherits KuduTable. When the target table is a
Kudu table, this patch changed the allowed instance from ImpalaKuduTable
to KuduTable so that CTAS statement could be supported afterwards.

Testing:
 - Added various q tests to verify Hive's behavior regarding creating
   and dropping Kudu tables.

Change-Id: I52a128d19403541be25001d9f32e1ac2272a0d2f