CDPD-29140: Change how fromJavaClassName() is invoked after IMPALA-10797
In IMPALA-10797, the signature of HdfsFileFormat#fromJavaClassName() was
changed. Since this method was called during the instantiation of
ImpalaResultLocation(), we need to modify the constructor of
ImpalaResultLocation accordingly.

Specifically, this patch adds an additional input argument for 'serDe'
added in IMPALA-10797 when fromJavaClassName() is used so that Hive
could be compiled with Impala.

Testing:
 - Verified in FENG's pre-commit tests that Hive could be compiled with
   Impala without any failed test after applying this patch.

Change-Id: I56c2abea96449b8eee7df1bd9c73199e11ff82be