DWX-8433: Added support for NullIf function.
The NullIf function isn't directly supported by Impala. Both Hive
and Impala treat the NullIf function as a special case of the
"case" function. So code has been added to support this.

In addition, a bug was fixed in the current code.

The getCommonDecimalInputs method is used to determine all
decimal return values. The wrong input parameters were being
examined which is now fixed.

Change-Id: Ida61ac5cc9202aeed9bd7247776877ab4f5c9ca1