CDPD-28579: Create Materialized Views uses incorrect partitioning keys
When a ctas or materialized view contains partitions, the created table places
the partition columns at the end. The insert/select statement therefore
also needs the select portion to place the partitioned columns at the end.

The Unified Analytics specific code now adds a Project node on top of
the previous root RelNode to reorder the columns.

Note: The Hive code handles this via the SemanticAnalyzer.saveViewDefinition
method. This code needs to exist for both CBO and non-CBO plans. Because of
this, the code that creates a Calcite node could not be made generic.
If we ever decide to desupport the non-CBO path, this should be refactored
(CDPD-30321).

Change-Id: I5f32979d00f6283b4ce64b94cd69c854fcd7e678