CDPD-35815: HIVE-25632: Remove unused code from ptest/ptest2 framework (Stamatis Zampetakis, reviewed by Zoltan Haindrich)
Ptest framework was deprecated when PTest2 was introduced, and the
latter is no longer used since it was superseded by HIVE-22942.

The code is more or less dead and keeping it in the repo leads to
maintenance overhead.

Drop also the following files since they are not used either and have
dependencies with ptests:
* jenkins-execute-build.sh
* jenkins-execute-hms-test.sh
* execute-test-on-lxc.sh

Closes #2736

This commit does not contain secrets.

(cherry picked from commit fc349e719f840b5b863e5d4d4a1b2a31449709e9)
Change-Id: I92466b63033f4719331390def3b6226a5e14dbe2