CDPD-40429: FENG: Support for Hive generic UDF in FENG.
This commit allows the Hive Generic UDFs to be visible
for the ImpalaFunctionResolver class.

Some of the resolvers had to be changed to create a dummy
signature rather than return null since a dummy signature
shows that the function can still be resolved by the Impala
function resolver.

The runtime environment is a bit tricky to solve. The Hive
functions will be in a jar file and the Impala runtime
environment may be on a separate machine. A Hive config
param "hive.impala.hive.exec.jar.hdfs.location" allows Hive
to send the function location to the Impala server at runtime.

Change-Id: Idfcbad2c8931e5e7612d29ae290b124ad1a316dd