HIVE-25675: Intermittent PSQLException when trying to connect to Postgres in tests
Sometimes when we try to connect to Postgres, the database is not yet
completely ready despite the fact that the respective port is open thus
leading to the following exception.

PSQLException: FATAL: the database system is starting up

1. Ensure the database is ready by checking a certain message appears
two times in the logs.
2. Remove the now unused logger.

(cherry picked from commit 206874646279c60e003f1d9745252045a710d5ec not
yet merged in apache/master)

Change-Id: I1a7cdb577f9120af8c4abc940458c20bbcd9158a