CDPD-31388: Do not advance estimated Impala time when RTT is short
We found that after IMPALA-11005 (part 2), there is a change in behavior
of clock_gettime(OsInfo::fast_clock(), &ts) that is called when the
Impala server invokes MonotonicStopWatch::Now() in
RuntimeProfile::EventSequence::Start() to validate the difference
between the current time on the Impala server and the estimated Impala
time (by HiveServer2) when HiveServer2 was submitting the query.

Recall that HiveServer2 has to fill in the field 'remote_submit_time' in
'TExecRequest' so that the information of query submission time will be
available to the Impala server. To do so, HiveServer2 sends a ping
request to the Impala HS2 server, which in turn calls
MonotonicStopWatch::Now() to retrieve the timestamp value from
clock_gettime(). After receving the response from Impala, HiveServer2
calculates the estimated Impala time by adding half of the roundtrip
time (RTT) to the timestamp value in the ping response.

Before IMPALA-11005 (part 2), we found that when 'clocksource' is
"kvm-clock" and 'clockid_t' is CLOCK_MONOTONIC_COARSE, even if RTT is
less than 4ms, the second call to clock_gettime() (to validate the
difference between the current Impala time and 'remote_submit_time')
would still advance the timestamp value by at least 4 ms.

However, after IMPALA-11005 (part 2), we observed that clock_gettime()
does not always advance the timestamp value in the second invocation
when RTT is less than 4ms. It is much more likely that clock_gettime()
fills in the same timestamp value as that in the ping response.

This change in behavior results in the Impala coordinator hitting a
DCHECK_GE that requires MonotonicStopWatch::Now() is at least -3ns later
than 'remote_submit_time' in 'TExecRequest' sent from HiveServer2 when
RTT between HiveServer2 and the Impala server is short, e.g., less than
4ms.

To avoid hitting the DCHECK_GE, we do not advance the estimated Impala
time when the mesaured RTT on the Hive side is less than
HIVE_IMPALA_MAX_TIMER_ERROR.

Change-Id: Iace522bc0c147d5fad92a52a96678063da302c35