CDPD-30961: Query 54 from tpcds fails due to bad cardinality check
The ImpalaCardinalityCheck RelNode wasn't coded correctly. In Hive, the
node is created from a HiveAggregate that does a count(*) and a
HiveFilter on top of that which runs the sq_count_check function
and checks on the cardinality.

In Impala, these two nodes translate to the ImpalaCardinalityCheckRel
node which creats a TupleDescriptor matching the input of the HiveFilter
and creates an output Expr object of a hardcoded BIGINT of 1. At execution
time, the Impala code is hardcoded to check that there is only one row
passing through it.

Change-Id: Id5bfe657d5b2a3dc712fe5112bef88f7134cce57