CDPD-20802: cast as ... format implementation for Unified Analytics
The "cast as ... format" function has been implemented.

Since Impala doesn't have a "cast_format" function, special processing
was added to ImpalaRexCall when this function is passed through Calcite.
Also, it was necessary to add function prototypes to the impala_scalars.json
file.

The "cast_format" function is now mapped to the SqlOperator.CAST type. A special
FunctionResolver is added to handle the analysis from AST into Calcite.

The "cast as varchar format" and "cast as char format" are not included in
this patch. There are added complications with these which will be handled
later.

Change-Id: I7e8e30c2e480728fe2a5b7dd83e7a207cf40201e