CDPD-25823: Add logging statement to facilitate debugging
We found that sometimes Impala on the branch of cdpd-master could not
successfully start the Hive Metastore due to the script bin/hive not
being able to parse $HADOOP_VERSION correctly using a regular
expression. This patch adds an additional statement to print out
$HADOOP_VERSION when it cannot be correctly parsed in bin/hive so that
it would be easier for us to identify the root cause.

Change-Id: Ie0facd8140c25a9a518504672fd16c4bf4eedf7c