CDPD-33633: Collect transaction ID for each cached table
We found that currently in some scenario the transaction ID associated
with a transactional materialized view (MV) could not be passed to
getPartitionsByNamesRequest() or getTableRequest() of
ImpalaHdfsPartitionLoader, resulting in an IllegalStateException being
thrown. This is due to the fact that the transaction ID associated with
the MV is not added to the map from which those 2 methods retrieve the
transaction ID of the given table, i.e., the MV in this case.

This patch fixes the bug by constructing the map in such a way that
those 2 methods could find the transaction ID of such an MV.

Testing:
 - Manually verified that after this patch, the IllegalStateException
   will not be thrown if we execute a statement based on which a
   transactional MV was created.

Change-Id: I3080a5ac6583ed3d768f6f903c3620f9b9651e31