CDPD-33826: Insert overwrite not showing rows for FENG
This is a regression for FENG caused by CDPD-6264. That improvement
tracks new dynamic partition directories as they are processed by
the Tez engine. This information isn't available when the Impala engine
is processing the insert overwrite.

The FileSinkDesc.dynPartitionValues should be initialized to null to
indicate that it should not be used.

Change-Id: Id3ac7372f91eda384d2f4320a950b375414dd83d