CDPD-33765 FENG: Insert values ignoring target schema
The type of the expression on an insert statement needs to
match the target schema for parquet. The problem shows up
when trying to insert a null or a decimal value into a table.
At parsing time, it is impossible to distinguish the type
so a CastExpr needed to be added onto the result expression.

This commit also fixes when the insert statement is given
target columns on a partitioned table and the partitioned
columns are not at the end of the target columns. This was
found while testing the Impala test insert_permutation.test.

Change-Id: I665f3c77d986ee5d1b72e7932d8798b2f07d3cef