CDPD-34362: DWX-16322: Unified Analytics: Remove case and decode function that returns varchar
There is a size check for varchars at compile time though that
was causing problems on the Impala side.

Since varchars are treated the same as strings in Impala backend,
there is no need to have a case function that returns a varchar.

The decode function is similar to case so the varchar version was
also removed for this function.

Change-Id: I475a953c07f99057fe2b9047a555089fdaba6377