CDPD-20814: Mod needs to use Impala arithmetic resolvers.
The mod function now goes through the Impala function resolver
method to determine the function to be used on the backend.

Also added "mod" and "div" to the list of operators that can be
used. Before this checkin, the "%" operator worked fine, but an
explicit "mod(a, b)" was causing a parsing error.

Change-Id: I292b534d31946f3d78b1d4ec84582675718de5b8