CDPD-35070: Unified Analytics: handle UDFs with explicit decimal types.
Decimals in C++ UDFs are defined explicitly rather than using a
wildcard. For instance, a function might be defined as:
     Decimal(9,0) my_func(Decimal(8,0), Decimal(7,0))
All built-in Impala functions use a wildcard and are defined like:
     Decimal(*,*) my_func(Decimal(*,*), Decimal(*,*))

One of the major changes involves changing ImpalaFunctionSignature:

ImpalaFunctionSignature creation is now divided into two distinct flavors:
  1) All functions that are stored in the database which are put into
     memory. These are created using the Impala defined datatypes.
  2) Querying the signatures by creating a dummy ImpalaFunctionSignature
     and checking for a match from the functions created by 1). The querying
     is done by using the Calcite datatypes which get changed internally
     into Impala datatypes that are queryable.

     Some changes had to be made to some query callers to ensure that the types
     passed in to ImpalaSignature creation are Calcite types.

Within the matching of ImpalaFunctionSignatures, there are also two distinct
flavors for Decimal types.
  1) Decimals that use wildcards in their variables. All built-in created
     functions use wildcards, which have a precision of -1. In order to query
     these, the query signature needs to change its decimal into a wildcard to
     match.
  2) Decimals that have explicit types. The only decimals that have explicit
     types are C++ UDFs created by users. In this case, the datatypes need
     to match both the precision and scale.

Some changes were also made on the querying side (ImpalaFunctionResolverImpl)
for the determination of the casted types for the parameters and the return type.
In the case where there are multiple decimal parameters, a common decimal type
must be determined must be used (e.g. the IN clause).

Change-Id: Id89ad8f6ebb0b3468a36f84a04031e83278adada