CDPD-34769: Insert into a partitioned table not working correctly.
The partitioned table grabs its columns from the Partition object.

The code has been rewritten so that there is a TargetHelper object
which organizes all the target table information in one place and initializes
relevant variables from the Hive compilation structures. Also added more
comments that clarifies the code.

The main work of this object is to handle result expressions from the select
and map these expressions into the target table. There are three major
cases that are handled here. For the examples, we will use a target table
with the following columns: C1 int, C2 string, C3 int, P1 int, P2 string

1) CTAS and Create MV statements. The following is an example of this type:
CREATE TABLE TARGET_TBL PARTITIONED BY (P1, P2) AS SELECT C1, C2, C3, P1, P2
In this case, the top level result expressions
determine the types of the columns in the target table. The nonpartitioned
columns are positioned in front of the partitioned columns.

2) The Insert statement has a targeted list clause. For example, the
insert statement looks like this:
INSERT INTO TBL (C3, P2, C1) VALUES (null, 'part val', 8);
In this case, the partitioned values do not have to be at the end of the
list. There may be columns missing from the target table which will
be filled in either by a default value or null. And there may be casting
of expressions that need to be done. The TINYINT 8 will need to be cast to
an INT, and the null will need to also be case to an INT. All partitions
in this case are dynamic.

3) The insert statement does not have a targeted list clause. For example,
the insert statement looks like this:
INSERT INTO TBL PARTITION(P1=3, P2) values (null, 'hello', 3);
In this case, the partitioned values are specified in the PARTITION clause,
though this partition clause is optional. The partitions can be declared
statically (P1=3) or dynamically (P2). If the PARTITION clause is left
out, the partitions will be found at the end of the SELECT (or VALUES)
columns. Like case 2), the types may need casting.

The code that places an optional sort node on top of the select statement
has been slightly modified to deal with this new object, but essentially
still does the same thing: It will try to add a sort clause if there are
dynamic partitions so the data inserted will be grouped together into the
target directory.

Change-Id: I0abed6a7422ebdc7697dbd384dfe47f4ab2f141f