CDPD-35043: UA: Avoid constant folding for UDFs
User defined functions were crashing HS2 during the constant folding
process. Since there is no way to tell whether UDFs carry state, constant
folding is disabled during optimization.

Added an "isStateful" parameter in the function details. When this is set
for a function, the constant folding will not be implemented.

Change-Id: I2dbcea9cf24261c0a16777626b2690a404e7b71f