CDPD-34918: Use the fully qualified tablename as the key instead of a table object during loading
The ImpalaTableLoader was previously maintaining a map of metastore
table object to the ImpalaHdfsTable object. This is not efficient since
a hashCode is computed on more than 30 attributes of the table object.
Further, it leads to cases where no match is found if even one attribute
has a slight difference.

This patch fixes the behavior by using the fully qualified name of the
table (dbname.tablename) as the key which should be sufficient for
correctness and is also more efficient. Tested by running the test
case supplied in CDPD-34918.

Change-Id: Ic69807e116840261ef97aea74a8594bf999cb533