CDPD-25823 (Part 2): Log related variables if $HADOOP_VERSION is empty
We found that in some scenarios $HADOOP_VERSION could evaluate to an
empty string, resulting in the script of hive not being able to
determine the Hadoop version.

This patch logs the following when $HADOOP_VERSION evaluates to an
empty string to determine the root cause of the issue.
1. The value of $HADOOP.
2. The output returned from executing "$HADOOP version".

To gather more information about the result of executing
"$HADOOP version" when the Hadoop version cannot be determined, we also
export HADOOP_SHELL_SCRIPT_DEBUG as true so that the related variables
could be output, e.g., HADOOP_OPTS.

Testing:
 - Manually verified that the logging statements added could output the
   variables we would like to collect.

Change-Id: Ieaea05436d8daa7868923cb7f6bb10595b3f56b4