DWX-10301: Generate SemanticException for DML operations on transactional tables when engine is Impala.
For Impala query engine, DML operations (INSERT/UPDATE/DELETE) are
disallowed on full acid transactional tables. UPDATE/DELETE are
disallowed on insert-only transactional tables. MERGE is a
combination of UPDATE and INSERT so it is also disallowed.
Previously, these operations would error out with an obscure
exception such as index out of bounds. These operations are
handled in this patch with a proper SemanticException.

The SemanticException is only generated if the ETL engine is
not set (this is controlled by an existing property
hive.etl.execution.engine).

Testing: negative tests were added for insert, update, delete, merge
for both full-acid and insert-only transactional tables.

Change-Id: I3c2430a709932d288227b0cd4c6d37f4dab37e68