CDPD-36578: Convert the keys of Impala's query options to lowercase
This patch fixed some unit tests against FENG in Hive's PTest that were
broken after CDPD-35472, where we explicitly converted the keys of
Impala's query options to uppercase letters when adding them to the
corresponding HiveConf.

Specifically, this was a bug because conflicting key-value pairs like
"impala.explain_level=VERBOSE" and "impala.EXPLAIN_LEVEL=MINIMAL" could
be both added and even if "impala.EXPLAIN_LEVEL=MINIMAL" was added later
than "impala.explain_level=VERBOSE", since
"impala.explain_level=VERBOSE" was not overridden, the value of the
query option sent to Impala could still be "VERBOSE". This was mainly
determined by the order in which we iterate over the key-value pairs in
the HiveConf.

Testing:
 - Verified that Hive's PTest passes after this patch.

Change-Id: If48a761a5c5c8b246b2dd3b5e53cd0a857a92ef0