CDPD-40020: Fix typo on the backport of HIVE-22510
While working on another Jira, I noticed a typo while backporting
HIVE-22510.

It only affects vectorization, it only affects when a VectorUDFAdaptor
is used, and it only affects when the inner column is a Decimal64 that
needs to be typecast to a Decimal.

I couldn't come up with a scenario that reproduces the problem.  The
TestVector flag removed the code path which generates Decimal64 columns.
And the only place I see the Adaptor used is within ANALYZE TABLE which
also doesn't need the Decimal64 to Decimal conversion.

But this should be pushed anyway since the downstream code does not
match the upstream code.

Change-Id: Ifd7473f15fdddff49a67577cef31c72d3bc3f4da