CDPD-29905: Move Impala loading exception message to debug.
The Impala jar file is not necessary for Hive to run, even though it is
part of the build. Some runtime environments do not copy over the Impala
buildtime environment. A message will still show that Impala could not
be loaded, but the noisy exception message will be only displayed in
logging debug mode.

Change-Id: I2625a127e391847cd791237e71ec323e19ff49fa