HIVE-26208: Exception in Vectorization with Decimal64 to Decimal casting (#3344)
The exception was happening because the code that casts the Decimal64 to
Decimal was not being added in filtered expressions.

Once this code was added, it caused a regression in check_constraint.q. The
reason for this was because we do not want to convert Decimal64 to Decimal
if the expression explicitly handles decimal64 types. A method was added
to these classes that will prevent the conversion in these cases.

Change-Id: I734f9a5e2023a0be54cd50b6a08959a0f76ef47f