CDPD-43839: Initialize TBackendGflags properly in FENG's q tests
This patch changes how we initialize TBackendGflags in FENG's q tests.

Specifically, before this patch, we attempted to initialize
TBackendGflags in ImpalaQueryHelperImpl#createTestQueryOptions() for
each query in a q file. After this patch, we only initialize
TBackendGflags for each q file.

In addition, this patch corrected how we initialized TBackendGflags.
More precisely, we need to know the value type of a key when setting up
the key in TBackendGflags. Otherwise, we could encounter a
ClassCastException when calling setFieldValue() to set the value of the
key.

Some TBackendGflags that have been removed from Impala were also removed
from data/conf/impala/hive-site.xml in this patch.

Testing:
 - Verified in a Jenkins job that both TestImpalaCliDriver and
   TestNegativeImpalaCliDriver succeeded when we had IMPALA-11469 on
   the Impala side.

Change-Id: I0bb3e1b25b96a3869e60e18b44f74eec9d1d0371