CDPD-43839: Initialize TBackendGflags properly in FENG's q tests
This patch changes how we initialize TBackendGflags in FENG's q tests.

Specifically, before this patch, we attempted to initialize
TBackendGflags in ImpalaQueryHelperImpl#createTestQueryOptions() for
each query in a q file. After this patch, we only initialize
TBackendGflags for each q file.

In addition, this patch corrected how we initialized TBackendGflags.
More precisely, we need to know the value type of a key when setting up
the key in TBackendGflags. Otherwise, we could encounter a
ClassCastException when calling setFieldValue() to set the value of the
key.

Some TBackendGflags that have been removed from Impala were also removed
from data/conf/impala/hive-site.xml in this patch.

Testing:
 - Verified in a Jenkins job that both TestImpalaCliDriver and
   TestNegativeImpalaCliDriver succeeded when we had IMPALA-11469 on
   the Impala side.

Note:
 - Due to RELENG-18563, we are not really able to reference the classes
   within the artifact of impala-frontend in the CDWH build even though
   this is not a problem in Hive's PTest and FENG's pre-commit tests. To
   overcome this limitation, we initialize Impala's backend
   configuration in TestImpalaCliDriver and TestNegativeImpalaCliDriver
   under hive-it-impala, which is less elegant. Once RELENG-18563 is
   resolved, we should try reapplying the more elegant patch at
   https://gerrit.sjc.cloudera.com/c/cdh/hive/+/193111.

Change-Id: Ia470bd92887ac30949a9ee379478c3cb2d0d422c