CDPD-44912: Type unification for UA: Rank datatypes now match.
The datatype returned for the "rank" function is a BIGINT in Impala
and an INT for Tez. In order to match this up, we need to put a cast
around the rank function when running in Tez.

The implementation of this was to add a new rule after all of the
optimization is complete. The function still resolves through the Impala
function resolver when running UA so the plan will contain a return
type of BIGINT. The new rule will find the rank method, change the
return type to an INT to match for Tez and place the cast around it.

Change-Id: Iecc9e5e5d1671bb3c222811059011f53d9bf3f1e