CDPD-45313: FENG: Cast to a date with format stopped working on Calcite upgrade.
The issue here was most likely because of the CAST SqlOperator.  I looked at
the RexBuilder.makeCast call which uses the SqlOperator and it only handles
one operand.  A second operand was tacked on for the format and was only used
at the end when the RexNode was converted back into an Impala Expr object. The
call was generated by RexBuilder.makeCall instead of makeCast which is why it
was able to be done this way.

Most likely the Calcite 1.21 was ignoring the extra operator but Calcite 1.25
was probably stripping the extra operator.

The way to fix this is that since we are not using CAST as Calcite expects,
we need to use the OTHER operator instead of CAST. Since the function name
is included in the RexNode, and since that is what is used when converting
to the Impala Expr, no other code needs to be changed.

Change-Id: I45752c1db3c98299b00d4e77751d45dbfb4d3b92