CDPD-46776: UA: Regression on boolean expression with an or clause
When CDPD-26890 was applied, it caused a regression on this
SQL statement:

select id, null and id = 0 from functional.alltypestiny where id < 2 or false

The addition of CDPD-26890 makes a call to FunctionInfo.isGenericUDF which is not
supported within the Impala code.  This method needs to return true for this case.

We now make it return true for all scalar functions.  We have a similar override
for aggregates.

Change-Id: Ie3805a04907974a08a811791d776983da9727f2c