CDPD-47178: UA query crashes impalad with join and one side has a constant
After running through CBO, the ImpalaJoinRel.getCanonical method ensures
that the left side of an equijoin conjunct corresponds to the left child.

The previous implementation worked fine when both sides had an input ref.
However, it was not working properly in the case where one side has a
constant.

The code now looks for an input ref on either side of the equijoin
conjunct. There should be at least one. It then calculates if the
equijoin expression needs swapping.

Change-Id: I1c6915609fbe1633483485aed436e14bc082c427