DWX-13945: Fix typo in Oracle schema upgrade
The index DBPRIVILEGEINDEX already exists and causes the schematool to fail.
Based on the postgres script, the intended name was DCPRIVILEGEINDEX, which matches the prefix of the new table.

Tested by running the below command to verify that this name is not yet used:

```
CREATE UNIQUE INDEX DCPRIVILEGEINDEX ON HIVE1.DC_PRIVS (AUTHORIZER,NAME,PRINCIPAL_NAME,PRINCIPAL_TYPE,DC_PRIV,GRANTOR,GRANTOR_TYPE);
```

Change-Id: Ie4087f076588bac21fb54e89ad71e529e144badf