DWX-14014: Remove hardcoded port and service from metastore health check
While ideally we should get a fix for CDPD-5246, this commit fixes the workaround which hardcodes the postgres port. This is needed since in private cloud the postrges DB can be external with a custom port.

During fixing, I noticed that the host is also wrong, in default DBC we should connect to `postgres-service-default-warehouse` since that is the DB actually used by HMS.

The change should be backward compatible since default values are set to the ones currently used.

Change-Id: I05ecc3c788c902c2eb028eac115ec5c0980e4246