CDPD-48234: Unified Analytics: Refactor refresh statement
This commit is a refactorization of the "refresh" command.

The two main architectural changes are that the Parser and the
SemanticAnalyzer are now pluggable classes.

The Parser plugin change starts in the Compiler object and calls
the appropriate parser based on the engine type.

The SemanticAnalyzer plugin entry point is located in the
SemanticAnalyzerFactory class. If the top level node returns true
for the "usesDefaultEngine" method in ASTNode, it will call into
the plugin module. In order to override this method, the plugin
creates ImpalaASTNode objects which derive from ASTNode objects.
In this commit, it only overrides the REFRESH ASTNode.

The Impala parser plugin calls two different parsers. First, it
calls the HiveParser as it did previously. If Hive cannot parse
the sql statement, it tries the Impala parser.

The framework needs to have an ASTNode in order to deal with things
like lineage. So if the Impala parser is used to parse the syntax,
an ASTNode tree still needs to be created. For refresh, we need to
keep the ASTNode snippets of the table tree and the partition tree.

Within the refresh semantic analyzer, there are a couple of changes.
The refresh semantic analyzer does not derive from SemanticAnalyzer
but instead copies the model for other non-query semantic analyzers
like the create function one which derives from BaseSemanticAnalyzer.
Like the create function analyzer, the refresh analyzer populates the
ReadEntity object for the core code to handle various tasks. The
semantic analyzer also creates the work and task objects directly
rather than relying on the CompilerDesc to create them, so some code
from ImpalaCompiler was no longer needed.

Authorization could theoretically be handled in the Hive core code
right now. However, the current iteration of the code within
the Operation2Privilege object (not changed in this commit) shows
that it allows all users to refresh a given table and pushes off the
authorization to Impala (tested with this commit). A future commit
could potentially allow the security portion to be pluginable as well
but that has been deferred with this commit.

Change-Id: I95b78eee72c43aed39ffc613bda2c217e43dd77c