CDPD-48236: Preserve the order of Impala's query options added via SET
Impala's query option should be case insensitive. Therefore, for the
same query option, a value set later should always override the value
set earlier regardless of letter case.

Before we are able to make Impala's TQueryOptions available to HiveConf
(CDPD-18048), as a workaround, this patch adds the field of
'lowercaseProperties' to HiveConf so that the order of Impala's query
options could be preserved.

Note that after this patch, if we would like to load Impala's query
options from an XML file, e.g., the one defined by 'hiveSiteURL', we
still need to make sure the key of each Impala's query option in the XML
file is in lowercase so that conflicting key-value pairs will not
interfere with each other. For instance,
("impala.explain_level", "MINIMAL") added via SET later could interfere
with ("impala.EXPLAIN_level", "VERBOSE") loaded from the XML file.
This could result in both key-value pairs being added to
'properties' of HiveConf, according to which HiveServer2 prepares
Impala's query options for Impala's backend.

Testing:
 - Added test cases to verify the issue is resolved with this patch.

Change-Id: I50a8ecd921bf5999f7f589b80f8b863f1b3ede13