CDPD-48534: UA: Refactor drop and compute statistics
This commit contains a refactoring of drop and compute statistics
on the Impala engine.

The one feature change included in this refactoring is that the Impala
syntax now works for these statements. The Impala syntax for computing
statistics is "COMPUTE STATS <tblname> ..." as opposed to
"ANALYZE TABLE <tblname> COMPUTE STATISTICS".

The commands are parsed through the Impala parser (in the impala-frontend
jar file). The fields are then parsed out of the Impala Stmt objects and
an ASTNode tree is generated from the Impala Stmt. The ASTNode
tree is needed in order to run through auditing and lineage and potentially
authorization in the future, which is handled in the ql code.

A special step needed to be added for compute statistics. We want to allow
the Hive compute statistics syntax to work the same as the Impala syntax.
When the Hive syntax succees parsing, we need to change the ASTNode tree
to match the Impala ASTNode tree.  This is done through the function
convertTreeIfNecessary.

Another change made is that the invalidate metadata task is now done
as a task generated in the semantic analyzer rather than being generated
in the ImpalaCompiler.  By placing code in the semantic analyzer, it
removed some command specific "if" statements in ImpalaCompiler, and a
bunch of code was deleted there.

An ImpalaFetchWork class was also created for commands such as these which
need the invalidate metadata. By instantiating this class in the semantic
analyzer, the ImpalaCompiler can determine if a Fetch Task is needed. This
also was done to eliminate adding "if" code everytime we add a new command
that needs to run invalidate metadata.

Change-Id: I312fea6de11ae8944d9137dd7f9c47a317e8a2c3