CDPD-48535: Unified Analytics: Accept Impala syntax for create function command.
Unified analytics  now accepts Impala syntax for the create function command.
The code created is similar to CDPD-48534  and CDPD-48234 which were
refactorizations for Impala syntax.

The reload function task is immediately called after the create
function command is sent to Impala to give the Hive server
visibility of the new function. A bug was fixed on the reload
function code where the session should not be initialized, opened,
or closed anywhere outside of the SessionManager.

Because create function syntax can exist in both Hive and Impala,
there is a config flag that allows the user to toggle between
the error messages.

Change-Id: I9b4b1c4e589ae337639b154b2f61411ad4df2b20