DWX-14026: Call GetOperationStatus for all UA queries.
After an execute is called to Impala on select queries, a call
to GetOperationStatus is made in a loop to see if results are ready
before the fetch call is made.

The reason for this is to prevent timeouts (though it does have an
issue that the fetch isn't retrieved immediately which should be
addressed in the future).

This code has been added to "compute stats" so that it does not
timeout on DWX.

Change-Id: I7264e6962ded27cceb50be03869c0daf616c5b90