CDPD-49641: Fix build after merging IMPALA-11745
Impala upstream was merged into Impala downstream which broke some
Unified Analytics code.  This patch fixes the code to match
the API changes.

IMPALA-11745 also added a new dependency. The Builtins loading
now depends on BackendConfig.INSTANCE. This is being loaded
when the jar is loaded, but there is no HiveConf currently
available. A quick hack was added here to allow progress of the
staging merge but this should be rewritten.

Change-Id: I6ae8587d10e6f4d3218debaccfff4a394d3e4cab