CDPD-50246: Invalidate statement needs to use correct db for table.
The invalidate metadata statement is run before compute stats. The statement
is created by grabbing the table and the appropriate database. The fix here
was that it needs to use the db from the compute stats statement if it is
provided there.

On top of this, the invalidate metadata statement also needs to run
synchronously, so this was added into this commit as well.

Change-Id: Icb285c404091f74c7ea215a0d41cd5501d156e9b